+++ vue-components/src/components/Button.vue
@@ -1,11 +1,14 @@
 <template>
-	<button :class="[ 'wikit', 'wikit-Button', `wikit-Button--${ type }` ]">
+	<button :class="[ 'wikit', 'wikit-Button', `wikit-Button--${ type }`, `wikit-Button--${ variant }` ]">
 		<slot />
 	</button>
 </template>
 
 <script lang="ts">
 import Vue from 'vue';
+import VueCompositionAPI, { defineComponent, onMounted } from '@vue/composition-api';
+
+Vue.use( VueCompositionAPI );
 
 /* eslint-disable no-trailing-spaces */
 /**
@@ -19,19 +22,49 @@ import Vue from 'vue';
  * https://bugzilla.mozilla.org/show_bug.cgi?id=1581369#c5
  */
 /* eslint-enable no-trailing-spaces */
-export default Vue.extend( {
+export default defineComponent( {
 	name: 'Button',
 	props: {
 		/**
 		 * The type of the button
+		 *
+		 * Allowed values: `neutral`, `progressive`, `destructive`
 		 */
 		type: {
 			type: String,
 			validator( value: string ): boolean {
-				return [ 'neutral', 'primaryProgressive', 'primaryDestructive' ].includes( value );
+				return [ 'neutral', 'progressive', 'destructive' ].includes( value );
 			},
 			default: 'neutral',
 		},
+		/**
+		 * The variant of the button
+		 *
+		 * Allowed values: `normal`, `primary`
+		 */
+		variant: {
+			type: String,
+			validator( value: string ): boolean {
+				return [ 'normal', 'primary' ].includes( value );
+			},
+			default: 'normal',
+		},
+	},
+	setup( props: {
+		type: 'neutral'|'progressive'|'destructive';
+		variant: 'normal'|'primary';
+	} ) {
+		onMounted( () => {
+			const supportedCombinations = {
+				normal: [ 'neutral' ],
+				primary: [ 'progressive', 'destructive' ],
+			};
+			if ( !supportedCombinations[ props.variant ].includes( props.type ) ) {
+				throw new Error(
+					`The combination of variant "${props.variant}" and type "${props.type}" is not yet supported!`,
+				);
+			}
+		} );
 	},
 } );
 </script>
@@ -61,7 +94,7 @@ $base: '.wikit-Button';
 		padding-block: $wikit-Button-large-padding-vertical;
 	}
 
-	&:not(:disabled) {
+	&#{$base}--normal {
 		&#{$base}--neutral {
 			color: $wikit-Button-normal-neutral-color;
 			background-color: $wikit-Button-normal-neutral-background-color;
@@ -78,7 +111,6 @@ $base: '.wikit-Button';
 				background-color: $wikit-Button-normal-neutral-active-background-color;
 				border-color: $wikit-Button-normal-neutral-active-border-color;
 			}
-
 			// A clicked button is both :active and :focused. Using :not(:active) to avoid mixing the two.
 			&:focus:not(:active) {
 				color: $wikit-Button-normal-neutral-focus-color;
@@ -87,8 +119,10 @@ $base: '.wikit-Button';
 				box-shadow: $wikit-Button-normal-neutral-focus-box-shadow;
 			}
 		}
+	}
 
-		&#{$base}--primaryProgressive {
+	&#{$base}--primary {
+		&#{$base}--progressive {
 			color: $wikit-Button-primary-color;
 			background-color: $wikit-Button-primary-progressive-background-color;
 			border-color: $wikit-Button-primary-progressive-border-color;
@@ -111,7 +145,7 @@ $base: '.wikit-Button';
 			}
 		}
 
-		&#{$base}--primaryDestructive {
+		&#{$base}--destructive {
 			color: $wikit-Button-primary-color;
 			background-color: $wikit-Button-primary-destructive-background-color;
 			border-color: $wikit-Button-primary-destructive-border-color;
@@ -135,11 +169,13 @@ $base: '.wikit-Button';
 		}
 	}
 
-	&:disabled {
+	&#{$base}--normal:disabled,
+	&#{$base}--primary:disabled {
 		color: $wikit-Button-normal-disabled-color;
 		background-color: $wikit-Button-normal-disabled-background-color;
 		border-color: $wikit-Button-normal-disabled-border-color;
 		cursor: default;
+		pointer-events: none;
 	}
 
 	// should ideally be taken care of by the globally applied style reset (ress)